Use a TIMESTAMP field type instead of the integer field to store the time.
Under the old schema, if a large number of log messages were written within
a short time frame (e.g. under a second), only the most recent log message
would actually persist.  In other words, the granuality of Unix time is too
poor to use for logging.

TIMESTAMP is (or should be) a standard SQL column type and is a more
efficient and correct means of storing time in an SQL database.


git-svn-id: http://svn.php.net/repository/pear/packages/Log/trunk@96534 c90b9560-bf6c-de11-be94-00142212c4b1