The SQL handler now uses sequences instead of timestamps to uniquely identify log entries.
We were using the NOW() function to generate quasi-unique timestamps, but
NOW() is now portable.  Instead, we generate timestamps using PHP's date()
function.  However, this results in a loss of timing granularity (times
are only accurate to the second), so we needed to add an 'id' column to
maintain a unique primary key.

This closes bug #20731.

I've also officially registered the dependency on the DB package.


git-svn-id: http://svn.php.net/repository/pear/packages/Log/trunk@116805 c90b9560-bf6c-de11-be94-00142212c4b1