The composite handler now opens child handlers lazily (on demand).
When logging from the composite level, if a unopened child will respond to a
log event, we attempt to open it immediately.  The composite handler's opened
state will be enabled as soon as the first child handler is successfully
opened.

We track an overall success state that indicates whether or not all of the
relevant child handlers were opened and successfully logged the event.  If one
handler fails, we still attempt any remaining children, but we consider the
overall result a failure.

No other changes regarding priorities or masking at the composite level have
been introduced, although they might be appropriate changes in the future if
backwards compatibility can be maintained.

The 'composite' unit test has been updated to reflect these changes.

Bug: 17785


git-svn-id: http://svn.php.net/repository/pear/packages/Log/trunk@303597 c90b9560-bf6c-de11-be94-00142212c4b1